完善发布页面功能

+++ blog-common/src/main/java/top/modelx/blog/common/table/entity/SysUser.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
  * </p>
  *
  * @author zhenhua
- * @since 2024-12-05
+ * @since 2024-12-28
  */
 @TableName("sys_user")
 public class SysUser implements Serializable {
@@ -22,7 +22,7 @@ public class SysUser implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     /**
      * 昵称
@@ -120,11 +120,11 @@ public class SysUser implements Serializable {
     @TableField("update_time")
     private LocalDateTime updateTime;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
